+++ src/main/java/com/bjut/MB/service/OrderService.java
@@ -21,7 +21,13 @@ public class OrderService {
     @Autowired
    private OrderDao orderDao;
 
-    public Map<String, String> addOrder(String orderNum,String process,String operater,String other,String ps){
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param process   工序名称
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
+    public Map<String, String> addOrder(String orderNum,String process){
         Map<String, String> map = new HashMap<String, String>();
         if(StringUtils.isBlank(orderNum)){
             map.put("code","2");
@@ -44,6 +50,15 @@ public class OrderService {
         return map;
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param process   工序名称
+     * @param operater  操作者
+     * @param other     其他内容说明
+     * @param ps        备注
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String, String> updateOrder(String orderNum,String process,String operater,String other,String ps){
         Map<String, String> map = new HashMap<String, String>();
         if(StringUtils.isBlank(orderNum)){
@@ -57,7 +72,7 @@ public class OrderService {
             return map;
         }
         try {
-            orderDao.updateItem(orderNum, process, operater, other, ps);
+            orderDao.updateItem(orderNum, operater, process, other, ps);
             map.put("code","1");
         }
         catch (Exception e){
@@ -67,12 +82,27 @@ public class OrderService {
         return map;
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个LIST集合
+     */
     public List<Order> selectOrder(String orderNum){
         return orderDao.selectAll(orderNum);
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return           返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String, String> deleteOrder(String orderNum){
         Map<String, String> map = new HashMap<String, String>();
+        if(StringUtils.isBlank(orderNum)){
+            map.put("code","2");
+            map.put("msg", "随工单编号不能为空！");
+            return map;
+        }
         try {
             orderDao.deleteAll(orderNum);
             map.put("code","1");
