+++ src/main/java/com/bjut/MB/service/PackService.java
@@ -22,7 +22,13 @@ public class PackService {
     @Autowired
     private PackDao packDao;
 
-    public Map<String,String> addPack(String orderNum, String process, String result, String operater){
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param process   名称
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
+    public Map<String,String> addPack(String orderNum, String process){
         Map<String, String> map = new HashMap<String, String>();
         if(StringUtils.isBlank(orderNum)){
             map.put("code","2");
@@ -45,7 +51,16 @@ public class PackService {
         return map;
     }
 
-    public Map<String, String> updatePack(String orderNum, String process, String result, String operater){
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param process   名称
+     * @param check     确认
+     * @param result    自检项目
+     * @param operater  包装人
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
+    public Map<String, String> updatePack(String orderNum,String process, String check, String result, String operater){
         Map<String, String> map = new HashMap<String, String>();
         if(StringUtils.isBlank(orderNum)){
             map.put("code","2");
@@ -58,7 +73,7 @@ public class PackService {
             return map;
         }
         try {
-            packDao.updateItem(result, operater, orderNum, process);
+            packDao.updateItem(check, result, operater, orderNum, process);
             map.put("code","1");
         }
         catch (Exception e){
@@ -68,12 +83,27 @@ public class PackService {
         return map;
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个LIST集合
+     */
     public List<Pack> selectPack(String orderNum){
         return packDao.selectAll(orderNum);
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String, String> deletePack(String orderNum){
         Map<String, String> map = new HashMap<String, String>();
+        if(StringUtils.isBlank(orderNum)){
+            map.put("code","2");
+            map.put("msg", "装箱记录单编号不能为空！");
+            return map;
+        }
         try {
             packDao.deleteAll(orderNum);
             map.put("code","1");
