+++ src/main/java/com/bjut/MB/service/PerformTestService.java
@@ -22,7 +22,13 @@ public class PerformTestService {
     @Autowired
     private PerformTestDao performTestDao;
 
-    public Map<String,String> addPerformTest(String orderNum, String process, String data, String result, String detectionDevice, String deviceType, String deviceNum, String ps){
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param process   序号
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
+    public Map<String,String> addPerformTest(String orderNum, String process){
         Map<String, String> map = new HashMap<String, String>();
         if(StringUtils.isBlank(orderNum)){
             map.put("code","2");
@@ -45,7 +51,16 @@ public class PerformTestService {
         return map;
     }
 
-    public Map<String, String> updatePerformTest(String orderNum, String process, String data, String result, String detectionDevice, String deviceType, String deviceNum, String ps){
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param process   序号
+     * @param data      实际数据
+     * @param result    结论
+     * @param ps        备注
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
+    public Map<String, String> updatePerformTest(String orderNum, String process, String data, String result, String ps){
         Map<String, String> map = new HashMap<String, String>();
         if(StringUtils.isBlank(orderNum)){
             map.put("code","2");
@@ -58,7 +73,7 @@ public class PerformTestService {
             return map;
         }
         try {
-            performTestDao.updateItem(orderNum, process, data, result, detectionDevice, deviceType, deviceNum, ps);
+            performTestDao.updateItem(orderNum, process, data, result, ps);
             map.put("code","1");
         }
         catch (Exception e){
@@ -68,12 +83,27 @@ public class PerformTestService {
         return map;
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个LIST集合
+     */
     public List<PerformTest> selectPerformTest(String orderNum){
         return performTestDao.selectAll(orderNum);
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String, String> deletePerformTest(String orderNum){
         Map<String, String> map = new HashMap<String, String>();
+        if(StringUtils.isBlank(orderNum)){
+            map.put("code","2");
+            map.put("msg", "性能要求检验单编号不能为空！");
+            return map;
+        }
         try {
             performTestDao.deleteAll(orderNum);
             map.put("code","1");
