+++ src/main/java/com/bjut/MB/service/RemadeSercice.java
@@ -23,6 +23,18 @@ public class RemadeSercice {
     @Autowired
     private RemadeDao remadeDao;
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param date      日期
+     * @param number    更改单号
+     * @param updateSoftware    软件更改内容简述
+     * @param updateHardware    硬件更改内容简述
+     * @param updateContent     结构更改内容简述
+     * @param updatePeople      更改人
+     * @param testPeople        检查人
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String,String> addRemade(String orderNum, Date date,String number,String updateSoftware,String updateHardware,
                                         String updateContent,String updatePeople,String testPeople){
         Map<String, String> map = new HashMap<String, String>();
@@ -47,6 +59,18 @@ public class RemadeSercice {
         return map;
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @param date      日期
+     * @param number    更改单号
+     * @param updateSoftware    软件更改内容简述
+     * @param updateHardware    硬件更改内容简述
+     * @param updateContent     结构更改内容简述
+     * @param updatePeople      更改人
+     * @param testPeople        检查人
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String, String> updateRemade(String orderNum, Date date, String number, String updateSoftware, String updateHardware,
                                             String updateContent, String updatePeople, String testPeople){
         Map<String, String> map = new HashMap<String, String>();
@@ -71,12 +95,27 @@ public class RemadeSercice {
         return map;
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个LIST集合
+     */
     public List<Remade> selectRemade(String orderNum){
         return remadeDao.selectAll(orderNum);
     }
 
+    /**
+     *
+     * @param orderNum  产品编号
+     * @return          返回一个map，key:code时，value为1则正常；为2说明参数有错，并把信息放到msg的key里；为0说明数据库操作出错
+     */
     public Map<String, String> deleteRemade(String orderNum){
         Map<String, String> map = new HashMap<String, String>();
+        if(StringUtils.isBlank(orderNum)){
+            map.put("code","2");
+            map.put("msg", "返工记录表编号不能为空！");
+            return map;
+        }
         try {
             remadeDao.deleteAll(orderNum);
             map.put("code","1");
